Made some edits to the decision_tree_learner function
Change to check if attrs is empty by using len(attrs)==0 since attrs is a np darrray and does not like to be checked by 'not attrs'

I also changed the argument when we make a subtree by doing a recursive call. Since we are going to be calling this function multiple times with the same attribute, I used np.setdiff1d method instead of attrs.remove list method. The np.setdiff1d does not modified the attrs array but just retruns a new attrs array with the set difference of attra and a (attrs-a)

I also added comments along with changing some variables for more readability.